Commit to doubles

+++ main.py
@@ -65,7 +65,7 @@ def train(dataloader,model,optimizer,loss_fn,true_func,par_gen,egrid):
     model.train()
     batches = 10000
     for batch, (D,P) in enumerate(dataloader):
-        D = D.float()
+        D = D.double()
         pred = model(P)
         loss = loss_fn(pred,D)
